datasheet/requirements.txt
reportlab==4.1.0
Pillow==10.4.0
pypdf==4.3.1